refactor: androidAppのnamespaceをソースパッケージに統一
- androidAppのnamespaceをcom.yaeyama.linercheckerに変更
- R/BuildConfigのimportを14ファイルでcom.yaeyama.linercheckerに置換
- ユニットテストをcom.yaeyama.linerchecker配下に移動しパッケージを変更、旧ディレクトリを削除
- dataモジュールのnamespaceをcom.yaeyama.linerchecker.dataに変更（マニフェストマージの重複解消）
- .serena/project.ymlにread_only_memory_patternsとline_endingを追加
