+++ androidApp/src/main/java/com/yaeyama/linerchecker/MainApplication.kt
@@ -2,7 +2,7 @@ package com.yaeyama.linerchecker
 
 import android.app.Application
 import com.google.firebase.crashlytics.FirebaseCrashlytics
-import com.ikmr.banbara23.yaeyama_liner_checker.BuildConfig
+import com.yaeyama.linerchecker.BuildConfig
 import com.yaeyama.linerchecker.di.appModule
 import com.yaeyama.linerchecker.di.dataModule
 import com.yaeyama.linerchecker.di.viewModelModule
