+++ androidApp/src/main/java/com/yaeyama/linerchecker/ui/dashboard/DashBoardPage.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.ikmr.banbara23.yaeyama_liner_checker.R
+import com.yaeyama.linerchecker.R
 import com.yaeyama.linerchecker.ui.dashboard.component.DashBoardHeader
 import com.yaeyama.linerchecker.ui.dashboard.component.DashBoardRow
 import com.yaeyama.linerchecker.ui.theme.YaimafuniAndroidTheme
