+++ androidApp/src/main/java/com/yaeyama/linerchecker/ui/dashboard/component/DashBoardAppBar.kt
@@ -10,8 +10,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import com.ikmr.banbara23.yaeyama_liner_checker.BuildConfig
-import com.ikmr.banbara23.yaeyama_liner_checker.R
+import com.yaeyama.linerchecker.BuildConfig
+import com.yaeyama.linerchecker.R
 import com.yaeyama.linerchecker.ui.theme.AppBackgroundColor
 import com.yaeyama.linerchecker.ui.theme.YaimafuniAndroidTheme
