+++ androidApp/src/main/java/com/yaeyama/linerchecker/ui/dashboard/component/DashBoardHeader.kt
@@ -15,7 +15,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.ikmr.banbara23.yaeyama_liner_checker.R
+import com.yaeyama.linerchecker.R
 import com.yaeyama.linerchecker.ui.theme.YaimafuniAndroidTheme
 
 @Composable
