+++ androidApp/src/main/java/com/yaeyama/linerchecker/ui/typhoon/list/TyphoonListTopAppBar.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
-import com.ikmr.banbara23.yaeyama_liner_checker.R
+import com.yaeyama.linerchecker.R
 import com.yaeyama.linerchecker.ui.common.PreviewBox
 
 @OptIn(ExperimentalMaterial3Api::class)
