+++ androidApp/src/main/java/com/yaeyama/linerchecker/ui/weather/compose/WeatherListItemCardHeader.kt
@@ -11,7 +11,7 @@ import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.ikmr.banbara23.yaeyama_liner_checker.R
+import com.yaeyama.linerchecker.R
 
 @Composable
 fun WeatherListItemCardHeader(title: String) {
